Changed JPA provider to EclipseLink instead of Hibernate due to licensing incompatibilities with EPL. Related issue #92 Signed-off-by: Adem KOSE <tr.ademkose@gmail.com>

+++ kuksa-appstore/src/main/java/org/eclipse/kuksa/appstore/service/AppCategoryService.java
@@ -122,7 +122,7 @@ public class AppCategoryService {
 	public List<String> getAllId() {
 
 		List<String> list = new ArrayList<>();
-		for (BigInteger component : appCategoryRepository.getAllId()) {
+		for (Long component : appCategoryRepository.getAllId()) {
 			list.add(component.toString());
 		}
 		return list;
